Extract YAML file/folder validators from validations.py
Second thematic extraction from modules/validations.py:

- modules/validations_yaml_files.py (~770 lines, new):
  26 functions covering the whole "user gave us a YAML location,
  is it valid Kometa YAML?" pipeline for Metadata / Collection /
  Overlay / Playlist file types.  Three layered tiers:
  1. Text validators (given YAML text, does it parse + have the
     required top-level mapping)
  2. Location validators (URL or filesystem path, fetch text +
     delegate to tier 1) plus folder-iteration variants
  3. Payload validators + Flask endpoint wrappers (public API,
     dispatch on file/folder/url/git/repo source type)

modules/validations.py: 1252 -> 595 (-657 / -52.5%).
Combined with PR #1519: 1828 -> 595 (-1233 / -67.5% since sprint start).

Uses the same re-export + shared-module pattern established in #1519.
Imports _resolve_managed_library_path and _saved_custom_repo_base
from modules.validations_shared (already extracted in #1519).

Also cleaned up 3 imports from validations.py that were no longer
used after the extraction (os, urllib.parse, ruamel.yaml.YAML).

No behavior change -- validations.py re-exports every moved name so
external callers (modules.library_file_entries, quickstart.py,
blueprints.validation_routes) continue to work unchanged.

AST-verified: all 26 moved functions are byte-for-byte identical to
origin/develop.
